requirements.txt
piazza-api==0.13.0
python-dotenv==1.0.0
anthropic==0.40.0
beautifulsoup4==4.12.3
rich==13.9.4

